storybook: 테마 프로바이더를 스토리북에 적용
